requirements.txt
diffusers
transformers
accelerate
safetensors
pillow